+++ web/package.json
@@ -9,6 +9,7 @@
   "scripts": {
     "build": "gulp build --production",
     "start": "node src/server",
+    "start-hot": "node webpack/server",
     "start-dev": "gulp",
     "test": "gulp test"
   },
@@ -97,5 +98,9 @@
     "karma-notify-reporter": "^0.1.1",
     "karma-sourcemap-loader": "^0.3.5",
     "karma-webpack": "^1.7.0"
+  },
+  "devDependencies": {
+    "nodemon": "^1.7.0",
+    "webpack-dev-server": "^1.11.0"
   }
 }
